// src/app/models/flight.ts
export interface Flight {
    id: number;
    code: string;
    airlineRefId: number;
    airlineName: any;
    fromAirportRefId: number;
    fromAirportName: any;
    toAirportRefId: number;
    toAirportName: any;
    address1: string;
    cityRefId: number;
    cityName: any;
    pincode: number;
    telephone1: string;
    email1: string;
}